move flatten log to pkg
